Add correct cycle times to branches; Make cycles u32

+++ src/nes/cpu/opcodes/bcc.rs
@@ -5,7 +5,7 @@ use super::branch::Branch;
 pub struct Bcc { }
 
 impl FromRelative for Bcc {
-    fn from_relative(cpu: &mut Cpu) -> usize {
+    fn from_relative(cpu: &mut Cpu) -> u32 {
         Branch::branch_on_true(cpu, |c| !c.SR.is_set(Flags::Carry));
 
         2
